Steven checked in some changes yesterday which fixed a great number of failing tests by setting SCONSFLAGS='-warn=no-python-version' in def run() in TestSCons.py yesterday, where previously it was done in def __init__(...).
This fixed most of the broken tests, but broke some of the working tests.
I readded the logic to def __init__

In addition I've added an extra arg to the __init__ for TestSCons.py
ignore_python_version (which defaults to true).
For those tests which need to be able to see the deprecation and/or
test the functionality of the deprecation.