Delegate linker Tool.exists() to CC/CXX Tool.exists().
Even for linking, need to respect CC/CXX specified by the user (issue#1723).
And when CC is specified but not CXX, assume the user knows there is nothing
to link with CXX, and delegate to CC Tool.exists() only. However, this somehow
should be synchronized with link.smart_link() to choose the correct linker.

--HG--
branch : linker-exists-when-compiler-exists