The _r in AIX xlc_r means reentrant, not relocatable.
It does not make any sense to use 'xlc' for CC and 'xlc_r' for SHCC, as the
'_r' does stand for 'reentrant' rather than 'relocatable' or similar.

Avoid 'egrep' to parse the lslpp output, it's easy enough within python.

Needs output streams of _subproc.dummyPopen to be iterable.

--HG--
branch : aix-xlc_r-is-for-threads